backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.1

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic[email]==2.10.6
pydantic-settings==2.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6

# Scheduling algorithm
ortools==9.14.6206
pandas==2.2.3
numpy==2.2.1

# AWS/Serverless support
mangum==0.17.0

# Testing and development
requests==2.32.3
pytest==8.3.4
httpx==0.28.1